CHeckpoint move JSON deserializers into class header files [skip ci]
