+++ engine/include/xjmusic/content/InstrumentAudio.h
@@ -1,7 +1,7 @@
 // Copyright (c) XJ Music Inc. (https://xjmusic.com) All Rights Reserved.
 
-#ifndef INSTRUMENT_AUDIO_H
-#define INSTRUMENT_AUDIO_H
+#ifndef XJMUSIC_INSTRUMENT_AUDIO_H
+#define XJMUSIC_INSTRUMENT_AUDIO_H
 
 #include <string>
 #include <utility>
@@ -28,6 +28,25 @@ namespace XJ {
     std::string tones;
   };
 
+  /**
+   * Parse a InstrumentAudio from a JSON object
+   * @param json  input
+   * @param entity  output
+   */
+  void from_json(const json &json, InstrumentAudio &entity) {
+    EntityUtils::setRequired(json, "id", entity.id);
+    EntityUtils::setRequired(json, "instrumentId", entity.instrumentId);
+    EntityUtils::setIfNotNull(json, "name", entity.name);
+    EntityUtils::setIfNotNull(json, "waveformKey", entity.waveformKey);
+    EntityUtils::setIfNotNull(json, "transientSeconds", entity.transientSeconds);
+    EntityUtils::setIfNotNull(json, "loopBeats", entity.loopBeats);
+    EntityUtils::setIfNotNull(json, "tempo", entity.tempo);
+    EntityUtils::setIfNotNull(json, "intensity", entity.intensity);
+    EntityUtils::setIfNotNull(json, "event", entity.event);
+    EntityUtils::setIfNotNull(json, "volume", entity.volume);
+    EntityUtils::setIfNotNull(json, "tones", entity.tones);
+  }
+
 }// namespace XJ
 
-#endif//INSTRUMENT_AUDIO_H
+#endif//XJMUSIC_INSTRUMENT_AUDIO_H
