+++ engine/include/xjmusic/content/Library.h
@@ -1,7 +1,7 @@
 // Copyright (c) XJ Music Inc. (https://xjmusic.com) All Rights Reserved.
 
-#ifndef LIBRARY_H
-#define LIBRARY_H
+#ifndef XJMUSIC_LIBRARY_H
+#define XJMUSIC_LIBRARY_H
 
 #include <string>
 
@@ -21,6 +21,19 @@ namespace XJ {
     long long updatedAt{EntityUtils::currentTimeMillis()};
   };
 
+  /**
+   * Parse a Library from a JSON object
+   * @param json  input
+   * @param entity  output
+   */
+  void from_json(const json &json, Library &entity) {
+    EntityUtils::setRequired(json, "id", entity.id);
+    EntityUtils::setRequired(json, "projectId", entity.projectId);
+    EntityUtils::setIfNotNull(json, "name", entity.name);
+    EntityUtils::setIfNotNull(json, "isDeleted", entity.isDeleted);
+    EntityUtils::setIfNotNull(json, "updatedAt", entity.updatedAt);
+  }
+
 }// namespace XJ
 
-#endif//LIBRARY_H
+#endif//XJMUSIC_LIBRARY_H
