+++ engine/include/xjmusic/content/ProgramSequence.h
@@ -1,7 +1,7 @@
 // Copyright (c) XJ Music Inc. (https://xjmusic.com) All Rights Reserved.
 
-#ifndef PROGRAM_SEQUENCE_H
-#define PROGRAM_SEQUENCE_H
+#ifndef XJMUSIC_PROGRAM_SEQUENCE_H
+#define XJMUSIC_PROGRAM_SEQUENCE_H
 
 #include <string>
 #include <utility>
@@ -23,6 +23,20 @@ namespace XJ {
     int total{};
   };
 
+  /**
+   * Parse a ProgramSequence from a JSON object
+   * @param json  input
+   * @param entity  output
+   */
+  void from_json(const json &json, ProgramSequence &entity) {
+    EntityUtils::setRequired(json, "id", entity.id);
+    EntityUtils::setRequired(json, "programId", entity.programId);
+    EntityUtils::setIfNotNull(json, "name", entity.name);
+    EntityUtils::setIfNotNull(json, "key", entity.key);
+    EntityUtils::setIfNotNull(json, "intensity", entity.intensity);
+    EntityUtils::setIfNotNull(json, "total", entity.total);
+  }
+
 }// namespace XJ
 
-#endif//PROGRAM_SEQUENCE_H
+#endif//XJMUSIC_PROGRAM_SEQUENCE_H
