+++ engine/include/xjmusic/content/ProgramSequenceBindingMeme.h
@@ -1,7 +1,7 @@
 // Copyright (c) XJ Music Inc. (https://xjmusic.com) All Rights Reserved.
 
-#ifndef PROGRAM_SEQUENCE_BINDING_MEME_H
-#define PROGRAM_SEQUENCE_BINDING_MEME_H
+#ifndef XJMUSIC_PROGRAM_SEQUENCE_BINDING_MEME_H
+#define XJMUSIC_PROGRAM_SEQUENCE_BINDING_MEME_H
 
 #include <string>
 #include <utility>
@@ -28,9 +28,20 @@ namespace XJ {
      */
     static std::set<std::string> getNames(const std::set<ProgramSequenceBindingMeme> &programSequenceBindingMemes);
 
-
   };
+
+  /**
+   * Parse a ProgramSequenceBindingMeme from a JSON object
+   * @param json  input
+   * @param entity  output
+   */
+  void from_json(const json &json, ProgramSequenceBindingMeme &entity) {
+    EntityUtils::setRequired(json, "id", entity.id);
+    EntityUtils::setRequired(json, "programId", entity.programId);
+    EntityUtils::setRequired(json, "programSequenceBindingId", entity.programSequenceBindingId);
+    EntityUtils::setIfNotNull(json, "name", entity.name);
+  }
 
 }// namespace XJ
 
-#endif//PROGRAM_SEQUENCE_BINDING_MEME_H
+#endif//XJMUSIC_PROGRAM_SEQUENCE_BINDING_MEME_H
