+++ engine/include/xjmusic/content/ProgramSequenceChord.h
@@ -1,7 +1,7 @@
 // Copyright (c) XJ Music Inc. (https://xjmusic.com) All Rights Reserved.
 
-#ifndef PROGRAM_SEQUENCE_CHORD_H
-#define PROGRAM_SEQUENCE_CHORD_H
+#ifndef XJMUSIC_PROGRAM_SEQUENCE_CHORD_H
+#define XJMUSIC_PROGRAM_SEQUENCE_CHORD_H
 
 #include <string>
 #include <utility>
@@ -22,6 +22,19 @@ namespace XJ {
     float position{};
   };
 
+  /**
+   * Parse a ProgramSequenceChord from a JSON object
+   * @param json  input
+   * @param entity  output
+   */
+  void from_json(const json &json, ProgramSequenceChord &entity) {
+    EntityUtils::setRequired(json, "id", entity.id);
+    EntityUtils::setRequired(json, "programId", entity.programId);
+    EntityUtils::setRequired(json, "programSequenceId", entity.programSequenceId);
+    EntityUtils::setIfNotNull(json, "name", entity.name);
+    EntityUtils::setIfNotNull(json, "position", entity.position);
+  }
+
 }// namespace XJ
 
-#endif//PROGRAM_SEQUENCE_CHORD_H
+#endif//XJMUSIC_PROGRAM_SEQUENCE_CHORD_H
