+++ engine/include/xjmusic/content/ProgramSequenceChordVoicing.h
@@ -1,7 +1,7 @@
 // Copyright (c) XJ Music Inc. (https://xjmusic.com) All Rights Reserved.
 
-#ifndef PROGRAM_SEQUENCE_CHORD_VOICING_H
-#define PROGRAM_SEQUENCE_CHORD_VOICING_H
+#ifndef XJMUSIC_PROGRAM_SEQUENCE_CHORD_VOICING_H
+#define XJMUSIC_PROGRAM_SEQUENCE_CHORD_VOICING_H
 
 #include <string>
 #include <utility>
@@ -22,6 +22,19 @@ namespace XJ {
     std::string notes;
   };
 
+  /**
+   * Parse a ProgramSequenceChordVoicing from a JSON object
+   * @param json  input
+   * @param entity  output
+   */
+  void from_json(const json &json, ProgramSequenceChordVoicing &entity) {
+    EntityUtils::setRequired(json, "id", entity.id);
+    EntityUtils::setRequired(json, "programId", entity.programId);
+    EntityUtils::setRequired(json, "programSequenceChordId", entity.programSequenceChordId);
+    EntityUtils::setRequired(json, "programVoiceId", entity.programVoiceId);
+    EntityUtils::setIfNotNull(json, "notes", entity.notes);
+  }
+
 }// namespace XJ
 
-#endif//PROGRAM_SEQUENCE_CHORD_VOICING_H
+#endif//XJMUSIC_PROGRAM_SEQUENCE_CHORD_VOICING_H
