+++ engine/include/xjmusic/content/ProgramSequencePattern.h
@@ -1,7 +1,7 @@
 // Copyright (c) XJ Music Inc. (https://xjmusic.com) All Rights Reserved.
 
-#ifndef PROGRAM_SEQUENCE_PATTERN_H
-#define PROGRAM_SEQUENCE_PATTERN_H
+#ifndef XJMUSIC_PROGRAM_SEQUENCE_PATTERN_H
+#define XJMUSIC_PROGRAM_SEQUENCE_PATTERN_H
 
 #include <string>
 #include <utility>
@@ -23,6 +23,20 @@ namespace XJ {
     int total{};
   };
 
+  /**
+   * Parse a ProgramSequencePattern from a JSON object
+   * @param json  input
+   * @param entity  output
+   */
+  void from_json(const json &json, ProgramSequencePattern &entity) {
+    EntityUtils::setRequired(json, "id", entity.id);
+    EntityUtils::setRequired(json, "programId", entity.programId);
+    EntityUtils::setRequired(json, "programSequenceId", entity.programSequenceId);
+    EntityUtils::setRequired(json, "programVoiceId", entity.programVoiceId);
+    EntityUtils::setIfNotNull(json, "name", entity.name);
+    EntityUtils::setIfNotNull(json, "total", entity.total);
+  }
+
 }// namespace XJ
 
-#endif//PROGRAM_SEQUENCE_PATTERN_H
+#endif//XJMUSIC_PROGRAM_SEQUENCE_PATTERN_H
