+++ engine/include/xjmusic/content/ProgramVoiceTrack.h
@@ -1,7 +1,7 @@
 // Copyright (c) XJ Music Inc. (https://xjmusic.com) All Rights Reserved.
 
-#ifndef PROGRAM_VOICE_TRACK_H
-#define PROGRAM_VOICE_TRACK_H
+#ifndef XJMUSIC_PROGRAM_VOICE_TRACK_H
+#define XJMUSIC_PROGRAM_VOICE_TRACK_H
 
 #include <string>
 #include <utility>
@@ -22,6 +22,19 @@ namespace XJ {
     float order{};
   };
 
+  /**
+   * Parse a ProgramVoiceTrack from a JSON object
+   * @param json  input
+   * @param entity  output
+   */
+  void from_json(const json &json, ProgramVoiceTrack &entity) {
+    EntityUtils::setRequired(json, "id", entity.id);
+    EntityUtils::setRequired(json, "programId", entity.programId);
+    EntityUtils::setRequired(json, "programVoiceId", entity.programVoiceId);
+    EntityUtils::setIfNotNull(json, "duration", entity.name);
+    EntityUtils::setIfNotNull(json, "tones", entity.order);
+  }
+
 }// namespace XJ
 
-#endif//PROGRAM_VOICE_TRACK_H
+#endif//XJMUSIC_PROGRAM_VOICE_TRACK_H
